replacing AAPI with Asian American (cont)
We decided to shift our language, rephrasing to Asian Americans (previously AAPI). Pacific Islander communities in North Carolina have not explicitly supported solidarities with the Asian American communities during the rise of anti-Asian violence, and we would like to acknowledge that with the language on this website. However, we hope that Pacific Islanders in North Carolina have established communities and support during this time!